+++ src/components/dumb.js
@@ -3,7 +3,7 @@ import {connect as connectToReduxStore } from 'react-redux';
 import {connect as connectToDefinitions} from '../behaviours/definitions';
 import {connect as connectToFieldHelpers} from '../behaviours/field';
 import {connect as connectSmartData} from '../behaviours/smart-data';
-import {fetchEntity} from '../actions';
+import {fetchEntity} from '../actions/entity';
 // Dumb components
 import Form from './form';
 import Field from './field';
